move reference field split to common and #197
* Moves references field splitting logic to common config so that we
can have multiple reference fields sharing it.

* Removes parent array, if the field is completely splitted and it's
the only remaining element in its parent
